improved variant creation is now moved to shopsys/framework
- product is set as main variant on variant creation instead of creating a duplicate
