+++ packages/framework/src/Model/Product/ProductVariantFacade.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Model\Product;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Exception;
 use Shopsys\FrameworkBundle\Component\Image\ImageFacade;
-use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPriceRecalculationScheduler;
 use Shopsys\FrameworkBundle\Model\Product\Recalculation\ProductRecalculationDispatcher;
 
 class ProductVariantFacade
@@ -18,7 +16,6 @@ class ProductVariantFacade
      * @param \Shopsys\FrameworkBundle\Model\Product\ProductDataFactoryInterface $productDataFactory
      * @param \Shopsys\FrameworkBundle\Component\Image\ImageFacade $imageFacade
      * @param \Shopsys\FrameworkBundle\Model\Product\ProductFactoryInterface $productFactory
-     * @param \Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPriceRecalculationScheduler $productPriceRecalculationScheduler
      * @param \Shopsys\FrameworkBundle\Model\Product\Recalculation\ProductRecalculationDispatcher $productRecalculationDispatcher
      */
     public function __construct(
@@ -27,7 +24,6 @@ class ProductVariantFacade
         protected readonly ProductDataFactoryInterface $productDataFactory,
         protected readonly ImageFacade $imageFacade,
         protected readonly ProductFactoryInterface $productFactory,
-        protected readonly ProductPriceRecalculationScheduler $productPriceRecalculationScheduler,
         protected readonly ProductRecalculationDispatcher $productRecalculationDispatcher,
     ) {
     }
@@ -37,31 +33,18 @@ class ProductVariantFacade
      * @param \Shopsys\FrameworkBundle\Model\Product\Product[] $variants
      * @return \Shopsys\FrameworkBundle\Model\Product\Product
      */
-    public function createVariant(Product $mainProduct, array $variants)
+    public function createVariant(Product $mainProduct, array $variants): Product
     {
-        $mainProduct->checkIsNotMainVariant();
+        $mainProduct->setAsMainVariant();
 
-        $mainVariantData = $this->productDataFactory->createFromProduct($mainProduct);
-        $mainVariant = $this->productFactory->createMainVariant($mainVariantData, $mainProduct, $variants);
-        $this->em->persist($mainVariant);
-
-        try {
-            $this->em->flush();
-            $this->productFacade->setAdditionalDataAfterCreate($mainVariant, $mainVariantData);
-            $this->imageFacade->copyImages($mainProduct, $mainVariant);
-        } catch (Exception $exception) {
-            $this->productPriceRecalculationScheduler->reset();
-
-            throw $exception;
+        foreach ($variants as $variant) {
+            $mainProduct->addVariant($variant);
         }
+        $this->em->flush();
 
-        // @todo after handling variants this may be simplified
-        $this->productRecalculationDispatcher->dispatchSingleProductId($mainVariant->getId());
+        // variants are recalculated automatically
+        $this->productRecalculationDispatcher->dispatchSingleProductId($mainProduct->getId());
 
-        foreach ($mainVariant->getVariants() as $variant) {
-            $this->productRecalculationDispatcher->dispatchSingleProductId($variant->getId());
-        }
-
-        return $mainVariant;
+        return $mainProduct;
     }
 }
